pyproject.toml
[project]
name = "langgraph-sqlite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.20",
    "langchain-core>=0.3.76",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
